+++ .prettierrc.js
@@ -1,7 +1,1 @@
-module.exports = {
-  singleQuote: true,
-  semi: false,
-  printWidth: 100,
-  tabWidth: 2,
-  trailingComma: 'es5',
-}
+module.exports = require('@cgr/project-config/prettier/base')