+++ apps/admin-web/src/context/AppProvider.tsx
@@ -1,7 +1,8 @@
-import { defaultTheme } from '@admin-web/config/theme'
-import { client } from '@admin-web/services/apollo'
 import { ApolloProvider } from '@apollo/client'
 import { ThemeProvider } from '@mui/material'
+
+import { defaultTheme } from '@admin-web/config/theme'
+import { client } from '@admin-web/services/apollo'
 
 interface AppProviderProps {
   children: React.ReactNode