+++ apps/admin-web/src/pages/_app.tsx
@@ -1,7 +1,8 @@
 import type { AppProps } from 'next/app'
+
+import { Layout } from '@admin-web/common/Layout'
 import { AppProvider } from '@admin-web/context/AppProvider'
 import '@admin-web/styles/globals.css'
-import { Layout } from '@admin-web/common/Layout'
 
 export default function App({ Component, pageProps }: AppProps) {
   return (