+++ apps/api/src/main.ts
@@ -6,35 +6,16 @@ import { Logger } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { NestFactory } from '@nestjs/core'
 import { NestExpressApplication } from '@nestjs/platform-express'
-import { WinstonModule } from 'nest-winston'
-import * as winston from 'winston'
 
 import cookieParser from 'cookie-parser'
 
 import { AppModule } from './app/app.module'
 import { Configuration, validateConfig } from './config/configuration'
+import { createLogger } from './util/logger'
 
 async function bootstrap() {
   const app = await NestFactory.create<NestExpressApplication>(AppModule, {
-    logger: WinstonModule.createLogger({
-      handleExceptions: true,
-      handleRejections: true,
-      defaultMeta: {
-        app: 'cgr-api',
-      },
-      transports: [
-        new winston.transports.Console({
-          format: winston.format.combine(
-            winston.format.colorize({ all: true }),
-            winston.format.simple()
-          ),
-        }),
-        new winston.transports.File({
-          filename: 'log/app.log',
-          format: winston.format.combine(winston.format.timestamp(), winston.format.json()),
-        }),
-      ],
-    }),
+    logger: createLogger(),
   })
 
   const configService = app.get<ConfigService<Configuration>>(ConfigService)