+++ packages/project-config/eslint/base.js
@@ -1,0 +1,46 @@
+// @ts-check
+
+/** @type {import('eslint').Linter.Config} */
+const config = {
+  root: true,
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint/eslint-plugin', 'unused-imports'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
+  ],
+  rules: {
+    '@typescript-eslint/no-empty-function': 'off',
+    'unused-imports/no-unused-imports': 'warn',
+    'unused-imports/no-unused-vars': [
+      'warn',
+      { vars: 'all', varsIgnorePattern: '^_', args: 'after-used', argsIgnorePattern: '^_' },
+    ],
+  },
+  ignorePatterns: [
+    '.next',
+    'dist',
+    'apps/api/src/graphql.ts',
+    'packages/codegen/src/generated/index.tsx',
+  ],
+  overrides: [
+    {
+      files: ['*.js', '*.cjs'],
+      rules: {
+        '@typescript-eslint/no-var-requires': 'off',
+        'no-undef': 'off',
+      },
+    },
+  ],
+  settings: {
+    react: {
+      version: '18.2.0',
+    },
+    next: {
+      rootDir: ['apps/*/'],
+    },
+  },
+}
+
+module.exports = config